==> scripts/02_eda.Rmd <==
---
title: "Exploratory data analysis"
output: html_document
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(
  echo = TRUE,
  message = FALSE,
  warning = FALSE
)
```

```{r}
library(tidyverse)
train <- read_csv("../data/raw/train.csv")
```

```{r}
train %>% 
  ggplot(aes(x = factor(MSSubClass), y = SalePrice)) +
  geom_boxplot()
```

```{r}
train %>% 
  ggplot(aes(x = LotFrontage, y = SalePrice)) +
  geom_point()
```

```{r}
table(train$Street)

train %>% 
  ggplot(aes(x = Street, y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$Alley)

train %>% 
  ggplot(aes(x = Alley, y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$LandContour)

train %>% 
  ggplot(aes(x = LandContour, y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$Utilities)

train %>% 
  ggplot(aes(x = Utilities, y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$LotConfig)

train %>% 
  ggplot(aes(x = LotConfig, y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$LandSlope)

train %>% 
  ggplot(aes(x = LandSlope, y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$Neighborhood)

train %>% 
  ggplot(aes(x = Neighborhood, y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$Condition1)

train %>% 
  ggplot(aes(x = Condition1, y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$Condition2)

train %>% 
  ggplot(aes(x = Condition2, y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$BldgType)

train %>% 
  ggplot(aes(x = BldgType, y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$OverallCond)

train %>% 
  ggplot(aes(x = factor(OverallCond), y = SalePrice)) +
  geom_boxplot()
```

```{r}
train %>% 
  ggplot(aes(x = YearBuilt, y = SalePrice)) +
  geom_point() +
  geom_smooth()
```

```{r}
lm_fit <- lm(SalePrice ~ YearBuilt, data = train)
rd <- residuals(lm_fit)
tibble(rd = rd, remod = train$YearRemodAdd - train$YearBuilt) %>% 
  ggplot(aes(x = remod, y = rd)) +
  geom_point() +
  geom_smooth()
```

```{r}
table(train$ExterCond)

train %>% 
  ggplot(aes(x = ExterCond, y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$BsmtFullBath)

train %>% 
  ggplot(aes(x = factor(BsmtFullBath), y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$BsmtHalfBath)

train %>% 
  ggplot(aes(x = factor(BsmtHalfBath), y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$Functional)

train %>% 
  ggplot(aes(x = Functional, y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$FireplaceQu)

train %>% 
  ggplot(aes(x = FireplaceQu, y = SalePrice)) +
  geom_boxplot()
```

```{r}
train %>% 
  ggplot(aes(x = YearBuilt, y = GarageYrBlt)) +
  geom_jitter()

train %>% 
  ggplot(aes(x = YearRemodAdd - YearBuilt, y = GarageYrBlt - YearBuilt)) +
  geom_jitter()
```

```{r}
lm_fit <- lm(SalePrice ~ YearBuilt, data = train)
rd <- residuals(lm_fit)
tibble(rd = rd, GarageRemod = train$GarageYrBlt - train$YearBuilt) %>% 
  ggplot(aes(x = GarageRemod, y = rd)) +
  geom_point() +
  geom_smooth(method = "lm")
```

```{r}
table(train$GarageQual)

train %>% 
  ggplot(aes(x = GarageQual, y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$GarageCond)

train %>% 
  ggplot(aes(x = GarageCond, y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$PoolQC)

train %>% 
  mutate(PoolQC = ifelse(is.na(PoolQC), "missing", "not-missing")) %>% 
  ggplot(aes(x = PoolQC, y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$Fence)

train %>% 
  group_by(Fence) %>% 
  summarize(AvgPrice = mean(SalePrice, na.rm = TRUE))

train %>% 
  ggplot(aes(x = Fence, y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$MiscFeature)

train %>% 
  group_by(MiscFeature) %>% 
  summarize(AvgPrice = mean(SalePrice, na.rm = TRUE))

train %>% 
  ggplot(aes(x = MiscFeature, y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$MiscVal)

train %>% 
  ggplot(aes(x = factor(MoSold), y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$MoSold)

train %>% 
  ggplot(aes(x = factor(MoSold), y = SalePrice)) +
  geom_boxplot()
```

```{r}
table(train$YrSold)

train %>% 
  ggplot(aes(x = factor(YrSold), y = SalePrice)) +
  geom_boxplot()
```
